examples/minimal-hydra-example/requirements.txt
tensorflow
hydra-core
wandb
